+++ lib/supra/stream.ex
@@ -6,25 +6,40 @@ defmodule Supra.Stream do
 
   def unfold_next_batch(query, cursor_finder, where_fun, last_field_value, opts) do
     repo = Keyword.get(opts, :repo) || raise(Supra.Error, "missing required option :repo")
+    batch_transform = Keyword.get(opts, :batch_transform)
+    preloads = Keyword.get(opts, :preload)
 
     case query_batch(
            repo,
            query,
            where_fun,
            last_field_value,
-           Keyword.get(opts, :batch_size, @batch_size),
-           Keyword.get(opts, :preload, [])
+           Keyword.get(opts, :batch_size, @batch_size)
          ) do
       [] ->
         nil
 
       batch ->
+        batch =
+          batch
+          |> then(fn batch ->
+            if preloads,
+              do: repo.preload(batch, preloads),
+              else: batch
+          end)
+          |> then(fn batch ->
+            if batch_transform,
+              do: batch_transform.(batch),
+              else: batch
+          end)
+
         last = List.last(batch)
+
         {batch, cursor_finder.(last)}
     end
   end
 
-  def query_batch(repo, query, where_fun, last_field_value, batch_size, preloads) do
+  def query_batch(repo, query, where_fun, last_field_value, batch_size) do
     where_clause = where_fun.(last_field_value)
 
     query
@@ -35,7 +50,6 @@ defmodule Supra.Stream do
     end)
     |> Ecto.Query.limit(^batch_size)
     |> repo.all()
-    |> repo.preload(preloads)
   end
 
   def where_next_batch(field, :asc, last_value),